FIx custom tokenizers which change token length
FMDBTokenizerNext() was assuming that the token remained the same length as currentRange suggests, but documentation says this is the range of inputString being analyzed.  The existing implementation leads to junk characters being fed to sql as part of the token if the token is shortened.

The safest thing to do is to get the length of cursor->tokenString instead of using currentRange here.  (Alternatively, the library's user could be required to update currentRange.)